Refactor addReviewers function and add helper functions
This commit refactors the addReviewers function in content.js to improve readability and maintainability. It introduces two new helper functions: closeReviewerModal and verifyReviewers.

The closeReviewerModal function is responsible for closing the modal after adding the last reviewer.

The verifyReviewers function is used to verify that all reviewers have been successfully added. It checks the list of added reviewers against the provided usernames and alerts the user if any reviewers were not added.